+++ automations/decodex/automations.toml
@@ -66,6 +66,13 @@ required_paths = [
 	"apps/radar/src/lib.rs",
 	"automations/decodex/README.md",
 	"automations/decodex/automations.toml",
+	"automations/decodex/scripts/config/automation_eval/__init__.py",
+	"automations/decodex/scripts/config/automation_eval/constants.py",
+	"automations/decodex/scripts/config/automation_eval/evaluation.py",
+	"automations/decodex/scripts/config/automation_eval/io.py",
+	"automations/decodex/scripts/config/automation_eval/model.py",
+	"automations/decodex/scripts/config/automation_eval/render.py",
+	"automations/decodex/scripts/config/automation_eval/validators.py",
 	"automations/decodex/scripts/config/evaluate_automations.py",
 	"automations/decodex/scripts/config/sync_automations.py",
 	"automations/radar/automations.toml",
